Add files via upload
How to create a 4 way traffic light system
This repository contains a tutorial on how to create a 4 way traffic light system using transistors, a 9V battery, capacitors, resistors, and LED bulbs. You will learn how to design and build a simple circuit that can control the traffic lights at a four-way intersection. You will also learn how to use a breadboard, a multimeter, and a soldering iron to assemble and test your circuit.

Why this tutorial?
This tutorial is designed for beginners who want to learn how to use basic electronic components and tools to create a fun and practical project. You will learn the basics of how transistors, capacitors, and resistors work, as well as how to use them to create a timer and a switch. You will also learn how to use LED bulbs to simulate the traffic lights and how to connect them to the circuit.

How to get started?
To get started with this tutorial, you will need to have some basic knowledge of electricity and circuits. You will also need to have the following materials and tools:

A breadboard
A 9V battery and a battery clip
Four transistors (2N2222 or equivalent)
Four capacitors (100 uF or equivalent)
Eight resistors (1 kΩ or equivalent)
Twelve LED bulbs (three red, three yellow, three green, and three white)
Some jumper wires
A multimeter
A soldering iron and some solder

Where to get help?
If you encounter any problems or have any questions while following this tutorial, you can reach out to me via email or open an issue on GitHub. I will try to help you as soon as possible.

Who maintains and contributes to this tutorial?
This tutorial is maintained and updated by me, Nethru Randev Wickramasekara, an electronics hobbyist and enthusiast. I created this tutorial to share my passion and knowledge of electronics with others. If you find this tutorial helpful, please consider giving it a star on GitHub or sharing it with your friends. If you have any feedback or suggestions, please feel free to contact me or submit a pull request. I appreciate your support and contribution.